moved the client list into separate components and add styles
