+++ src/components/ClientId.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { Box, CircularProgress, Typography } from '@mui/material';
+import { Box, CircularProgress, Divider, Typography } from '@mui/material';
 import PostService from '../api/PostServise';
 import { useFetching } from '../hooks/useFetching';
 import { getClientId } from '../store/clients/actions';
@@ -10,7 +10,7 @@ interface Props {
   clientID: number | null;
 }
 
-export const ClientsId: FC<Props> = ({ clientID }) => {
+export const ClientId: FC<Props> = ({ clientID }) => {
   const dispatch = useDispatch();
   const client = useSelector((state: State) => state.clients.currentClient);
 
@@ -27,14 +27,21 @@ export const ClientsId: FC<Props> = ({ clientID }) => {
   if (!client) return null;
   return (
     <div>
-      <Typography>{clientID}</Typography>
-      <Typography>
-        {client.general.firstName} {client.general.lastName} <br />
-        {'Job:'} {client.job.company} {client.job.title} <br />
-        {'Address:'}
-        {client.address.country} {client.address.city} {client.address.street} {client.address.zipCode}
-      </Typography>
-
+      <div>
+        <Typography>
+          Name:
+          {client.general.firstName} {client.general.lastName}
+        </Typography>
+        <Divider />
+        <Typography>
+          Job: Company: {client.job.company} <br /> Title: {client.job.title} <br />
+        </Typography>
+        <Typography>
+          Address: Country:
+          {client.address.country} <br /> City: {client.address.city} <br /> Street: {client.address.street}
+          <br /> Zip code: {client.address.zipCode}
+        </Typography>
+      </div>
       {clientIdError && <h1 className="error">Произошла ошибка {clientIdError}</h1>}
       {isClientIdLoading && (
         <Box sx={{ display: 'flex' }}>
